@@ -3,17 +3,16 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@testing-library/jest-dom": "^5.16.1",
-    "@testing-library/react": "^12.1.2",
-    "@testing-library/user-event": "^7.2.1",
-    "http-proxy-middleware": "^1.0.4",
-    "nanoid": "^3.1.30",
+    "@testing-library/jest-dom": "^5.16.4",
+    "@testing-library/react": "^13.2.0",
+    "@testing-library/user-event": "^14.1.1",
+    "nanoid": "^3.3.4",
     "noty": "^3.2.0-beta",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-multi-carousel": "^2.5.5",
-    "react-router-dom": "^5.1.2",
-    "react-scripts": "^4.0.3"
+    "react": "^18.1.0",
+    "react-dom": "^18.1.0",
+    "react-multi-carousel": "^2.8.0",
+    "react-router-dom": "^6.3.0",
+    "react-scripts": "^5.0.1"
   },
   "scripts": {
     "start": "cross-env PORT=3001 react-scripts start",
